// packages/metajob-theme/src/blocks/resume-block/types.ts
//Updated resume types
export type ResumeFormProps = {
   name: ""
   tagline: ""
   qualification: ""
   experienceTime: ""
   language: ""
   salaryType: ""
   salary: number
   category: number
   about: string
   education: {
      title: string
      institution: string
      startDate: string
      endDate: string
      description: string
   }[]
   experience: {
      title: string
      institution: string
      startDate: string
      endDate: string
      description: string
   }[]
   portfolio: {
      title: string
      description: string
      link: {
         link: string
         label: string
         target: string
         type: string
      }
      // image: string;
   }[]
   contact: {
      friendlyAddress: string
      location: string
   }
   user?: any
}

export type IJobCategory = {
   id: number
   attributes: {
      title: string
   }
}